pyproject.toml
[project]
name = "marimo-claude"
version = "0.1.0"
description = "MCP server bridging Marimo notebooks and Claude — edit, inspect, and visualize notebooks collaboratively"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
authors = [{ name = "Bjoern Bethge" }]
keywords = ["marimo", "claude", "mcp", "notebook", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "marimo[mcp,recommended]>=0.21.1",
    "mcp-ui-server>=1.0.0",
]

[project.scripts]
marimo-claude = "marimo_claude.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/marimo_claude"]
